extract components for a better architecture of WWW
